% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_db_import_previous.R
\name{bill_db_import_previous}
\alias{bill_db_import_previous}
\title{Import historical legislative data into database}
\usage{
bill_db_import_previous(db_path, previous_bills_df)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} containing the path where the bill database is located.}

\item{previous_bills_df}{A dataframe or tibble. This will usually be created using \code{get_bill_table} and later loaded with a simple \code{read_csv} command.}
}
\value{
Nothing explicitly except for a success, skip, or error message. Inserts data from \code{previous_bills_df} into bill database created using \code{bill_db_init}.
}
\description{
Import historical legislative data into database
}
\examples{
db_path <- here::here("legislation.db")
session_url <- "https://www.legislature.mi.gov/Search/ExecuteSearch?sessions=2025-2026&docTypes=House\%20Bill,Senate\%20Bill"
previous_bills_df <- get_bill_table(session_url)
bill_db_import_previous(db_path = db_path, previous_bills_df = previous_bills_df)
# If successful, returns "Historical data import completed successfully."
# If db_check_history(db_path) is TRUE, returns "Historical bill data has already been imported. Skipping import."
# If unsuccessful, returns "Error importing historical bill data:" and the associated error message.
}
