% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_db_init.R
\name{bill_db_init}
\alias{bill_db_init}
\title{Initialize local database for storing bills}
\usage{
bill_db_init(db_path)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} that contains a path at which to create a database object. The preferred approach is to set the \code{db_path} beforehand using the \code{here} package (see Examples below). The database will be created if it does not already exist at the specified path; it will also create the \code{bills} and \code{metadata} tables for storing the data.}
}
\value{
Nothing explicitly; creates a database object at the specified path.
}
\description{
Initialize local database for storing bills
}
\examples{
bill_db_init(db_path = here::here("legislation.db"))
}
