% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_db_main.R
\name{bill_db_main}
\alias{bill_db_main}
\title{Main function to maintain legislation database}
\usage{
bill_db_main(db_path, previous_bills_df = NULL, current_bills_df = NULL)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} containing the path where the bill database is located.}

\item{previous_bills_df}{A dataframe or tibble, typically created using \code{get_bill_table} or called using \code{read_csv}, containing all historical bill data.}

\item{current_bills_df}{A dataframe or tibble, typically created using \code{get_bill_table} or called using \code{read_csv}, containing all current session bill data.}
}
\value{
Nothing explicitly; inserts and/or updates information in the database.
}
\description{
Main function to maintain legislation database
}
\examples{
db_path <- here::here("legislation.db")
bill_db_main(db_path = db_path)  # This will refresh the database to make it current, because data should already be living in a folder!
}
