% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_db_query.R
\name{bill_db_query}
\alias{bill_db_query}
\title{Flexibly query a legislation database}
\usage{
bill_db_query(db_path, session_years = NULL, current_session_only = FALSE)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} containing the path where the bill database is located.}

\item{session_years}{A character vector of \code{length = 1} containing a valid session-years string, such as "2025-2026".}

\item{current_session_only}{Set to \code{FALSE} by default; if changed to \code{TRUE}, returns only the current session's bills.}
}
\value{
A tibble.
}
\description{
Flexibly query a legislation database
}
\examples{
db_path <- here::here("legislation.db")
bill_db_query(db_path = db_path, session_years = "2015-2016", current_session_only = FALSE)
# Returns a tibble with data on all bills from the 2015-2016 legislative session.
}
