% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_db_update_current.R
\name{bill_db_update_current}
\alias{bill_db_update_current}
\title{Update current session legislation data in database}
\usage{
bill_db_update_current(db_path, current_bills_df)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} containing the path where the bill database is located.}

\item{current_bills_df}{A dataframe or tibble. This will usually be created using \code{get_bill_table} and later loaded with a simple \code{read_csv} command.}
}
\value{
Nothing explicitly except for a success or error message. Inserts data from \code{current_bills_df} into bill database created using \code{bill_db_init}.
}
\description{
Update current session legislation data in database
}
\examples{
db_path <- here::here("legislation.db")
session_url <- "https://www.legislature.mi.gov/Search/ExecuteSearch?sessions=2025-2026&docTypes=House\%20Bill,Senate\%20Bill"
current_bills_df <- get_bill_table(session_url)
bill_db_update_current(db_path = db_path, current_bills_df = current_bills_df)
# If successful, returns nothing.
# If unsuccessful, returns "Error updating current session data:" and the associated error message.
}
