% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check_history.R
\name{db_check_history}
\alias{db_check_history}
\title{Check for historical bill data updates}
\usage{
db_check_history(db_path)
}
\arguments{
\item{db_path}{A character vector of \code{length = 1} that contains the path location of a database object created by \code{bill_db_init}.}
}
\value{
\code{TRUE} if historical bill data has already been imported via \code{bill_db_import_previous}.
}
\description{
Check for historical bill data updates
}
\examples{
db_path <- here::here("legislation.db")
db_check_history(db_path = db_path)
# TRUE
}
