% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_bill_data.R
\name{fetch_bill_data}
\alias{fetch_bill_data}
\title{Compile Michigan Legislature data}
\usage{
fetch_bill_data(
  raw_dir = here::here("data", "raw"),
  clean_dir = here::here("data", "clean")
)
}
\arguments{
\item{raw_dir}{A character vector of \code{length = 1} that contains a path where raw data will be stored.}

\item{clean_dir}{A character vector of \code{length = 1} that contains a path where clean data will be stored.}
}
\value{
Nothing explicitly, but does create data directories and store flat files, which can be used independently or loaded into a legislation database created with \code{bill_db_init}.
}
\description{
Compile Michigan Legislature data
}
