% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bill_table.R
\name{get_bill_table}
\alias{get_bill_table}
\title{Extract a bill table from a MI Legislature session URL}
\usage{
get_bill_table(session_url)
}
\arguments{
\item{session_url}{A character vector with one element, which is a properly-formatted URL string.}
}
\value{
A tibble.
}
\description{
Extract a bill table from a MI Legislature session URL
}
\examples{
session_url <- "https://www.legislature.mi.gov/Search/ExecuteSearch?sessions=2025-2026&docTypes=House\%20Bill,Senate\%20Bill"
get_bill_table(session_url)
}
