% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_area.R
\name{get_doc_area}
\alias{get_doc_area}
\title{Extract document subject area categories}
\usage{
get_doc_area(doc_descriptions)
}
\arguments{
\item{doc_descriptions}{A character vector of document descriptions, typically as part of a tibble produced by \code{get_bill_table}.}
}
\value{
A character vector containing the major and minor subject area categories as "Major: minor".
}
\description{
Extract document subject area categories
}
\examples{
doc_description <- "Transportation: funds; funding for sound wall projects; provide for. Amends 1951 PA 51 (MCL 247.651 - 247.675) by adding sec. 14b."
get_doc_area(doc_descriptions = doc_description)
# "Transportation: funds"
}
