% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_id.R
\name{get_doc_id}
\alias{get_doc_id}
\title{Rearrange a document name/title into an ID}
\usage{
get_doc_id(doc_names)
}
\arguments{
\item{doc_names}{A character vector of legislative document names/titles, such as "HB 4001 of 2025".}
}
\value{
A character vector.
}
\description{
Rearrange a document name/title into an ID
}
\examples{
doc_name <- "HB 4001 of 2025"
get_doc_id(doc_names = doc_name)
doc_names <- c("HB 4001 of 2025", "SB 1 of 2025", "PA 127 of 2025")
get_doc_id(doc_names = doc_names)
}
