% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_session.R
\name{get_doc_session}
\alias{get_doc_session}
\title{Reformat legislative document session year string}
\usage{
get_doc_session(doc_names)
}
\arguments{
\item{doc_names}{A character vector of document names, such as "HB 4001 of 2025".}
}
\value{
A character vector of properly-formatted, two-year legislative session IDs, formatted as "2025-2026". Adjusts for bills with even-number years in their name to the correct "odd-even" structure.
}
\description{
Reformat legislative document session year string
}
\examples{
doc_name <- "HB 4001 of 2025"
get_doc_session(doc_names = doc_name)
# "2025-2026"
doc_names <- c("HB 4001 of 2025", "SB 1 of 2012", "PA 127 of 2018")
get_doc_session(doc_names = doc_names)
# "2025-2026" "2011-2012" "2017-2018")
}
