% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_url.R
\name{get_doc_url}
\alias{get_doc_url}
\title{Create MI Legislature document URL from document name}
\usage{
get_doc_url(doc_names)
}
\arguments{
\item{doc_names}{A character vector of legislative document names, such as "HB 4001 of 2025".}
}
\value{
A character vector of properly-formatted MI Legislature URLs for the given \code{doc_names}.
}
\description{
Create MI Legislature document URL from document name
}
\examples{
doc_name <- "HB 4001 of 2025"
get_doc_url(doc_names = doc_name)
# "https://www.legislature.mi.gov/Bills/Bill?ObjectName=2025-HB-4001"
}
