% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mileg-package.R
\docType{package}
\name{mileg-package}
\alias{mileg}
\alias{mileg-package}
\title{mileg: Compile and Analyze Michigan Legislative Data}
\description{
Tools for compiling and analyzing data on the Michigan Legislature.
}
\author{
\strong{Maintainer}: Aaron Meek \email{aaronameek@gmail.com}

}
\keyword{internal}
