% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bill_table.R
\name{process_bill_table}
\alias{process_bill_table}
\title{Process a table of bills}
\usage{
process_bill_table(raw_bills_table)
}
\arguments{
\item{raw_bills_table}{A tibble, typically provided by \code{get_bill_table}.}
}
\value{
A tibble.
}
\description{
Process a table of bills
}
\examples{
su <- "https://www.legislature.mi.gov/Search/ExecuteSearch?sessions=2025-2026&docTypes=House\%20Bill,Senate\%20Bill"
bt <- get_bill_table(su)
process_bill_table(bt)
}
